requirements.txt
streamlit
openai
requests
python-dotenv
pandas
Pillow 